back/pyproject.toml
[project]
name = "template_app"
version = "0.1.dev"
description = "FastAPI template with domain-driven architecture"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "python-dotenv>=1.1.0",
    "fastapi[standard]>=0.115.12",
    "sqlmodel>=0.0.24",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "pydantic>=2.0.0",
]

[dependency-groups]
dev = [
    "click>=8.1.8",
    "ruff>=0.8.0",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.1.0",
    "faker>=37.6.0",
    "mypy>=1.17.1",
]

[project.scripts]
init-db = "template_app.commands.init_db:main"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "FAST", "UP", "RUF", "B"]
ignore = ["E501", "RUF001", "RUF002", "RUF003"]

[tool.ruff.lint.per-file-ignores]
"**/database/*.py" = ["E712"]
"alembic/**.py" = ["ALL"]

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
filterwarnings = [
    "error", # treat all warnings as errors
    "ignore:unclosed.*:ResourceWarning", # ignore unclosed resource warnings during test cleanup
]
asyncio_mode = "auto"
addopts = "--disable-socket --allow-unix-socket --allow-hosts=127.0.0.1,localhost,::1"

[tool.mypy]
exclude = "(?x)(alembic/*)"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
